dependency injection

+++ app/src/main/java/com/android/fastfood/utils/CurrencyFormat.kt
@@ -4,7 +4,7 @@ import android.icu.text.NumberFormat
 import android.icu.util.Currency
 
 
-fun Double.currecyFormat():String{
+fun Int.currecyFormat():String{
     val price = NumberFormat.getCurrencyInstance()
     price.currency = Currency.getInstance("IDR")
     price.maximumFractionDigits = 0
